chore: add dev build with typechecking and autostart to exsample 00
